requirements.txt
PyQt6
Pillow
numpy
matplotlib
pyqtgraph